docker build trivially working
