+++ .gitignore
@@ -1,2 +1,10 @@
-*-venv/
 .idea/
+__pycache__
+a2c_ppo_acktr.egg-info
+RetroGym
+MarioWM/.*
+.eggs/
+.github/
+MarioWM/.gdb_history
+MarioWM/.gitignore
+MarioWM/.travis.yml
