+++ MarioWM/README.md
@@ -1,2 +1,2 @@
 # MarioWM
-Mario RL for testing weird-state exploitation 
+
