+++ MarioWM/a2c_ppo_acktr/agents/distributions.py
@@ -1,50 +1,42 @@
 import torch
 import torch.nn as nn
 
-from a2c_ppo_acktr.utils import AddBias, init
+from a2c_ppo_acktr.utils.train_utils import init, AddBias
 
 """
 Modify standard PyTorch distributions so they are compatible with this code.
 """
 
-#
-# Standardize distribution interfaces
-#
 
-# Categorical
 class FixedCategorical(torch.distributions.Categorical):
-    def sample(self):
+    def sample(self, **kwargs):
         return super().sample().unsqueeze(-1)
 
     def log_probs(self, actions):
-        return (
-            super()
-            .log_prob(actions.squeeze(-1))
-            .view(actions.size(0), -1)
-            .sum(-1)
-            .unsqueeze(-1)
-        )
+        return super().log_prob(actions.squeeze(-1)).view(actions.size(0), -1).sum(-1).unsqueeze(-1)
 
     def mode(self):
         return self.probs.argmax(dim=-1, keepdim=True)
 
 
-# Normal
 class FixedNormal(torch.distributions.Normal):
+    def enumerate_support(self, expand=True):
+        pass
+
     def log_probs(self, actions):
         return super().log_prob(actions).sum(-1, keepdim=True)
 
-    def entrop(self):
-        return super.entropy().sum(-1)
+    def entropy(self):
+        return super().entropy().sum(-1)
 
     def mode(self):
         return self.mean
 
 
-# Bernoulli
 class FixedBernoulli(torch.distributions.Bernoulli):
+
     def log_probs(self, actions):
-        return super.log_prob(actions).view(actions.size(0), -1).sum(-1).unsqueeze(-1)
+        return super().log_prob(actions).view(actions.size(0), -1).sum(-1).unsqueeze(-1)
 
     def entropy(self):
         return super().entropy().sum(-1)
@@ -57,11 +49,7 @@ class Categorical(nn.Module):
     def __init__(self, num_inputs, num_outputs):
         super(Categorical, self).__init__()
 
-        init_ = lambda m: init(
-            m,
-            nn.init.orthogonal_,
-            lambda x: nn.init.constant_(x, 0),
-            gain=0.01)
+        init_ = lambda m: init(m, nn.init.orthogonal_, lambda x: nn.init.constant_(x, 0), gain=0.01)
 
         self.linear = init_(nn.Linear(num_inputs, num_outputs))
 
@@ -70,9 +58,9 @@ class Categorical(nn.Module):
         return FixedCategorical(logits=x)
 
 
-class DiagGaussian(nn.Module):
+class FixedDiagGaussian(nn.Module):
     def __init__(self, num_inputs, num_outputs):
-        super(DiagGaussian, self).__init__()
+        super(FixedDiagGaussian, self).__init__()
 
         init_ = lambda m: init(m, nn.init.orthogonal_, lambda x: nn.init.
                                constant_(x, 0))
