--- MarioWM/a2c_ppo_acktr/algo/__init__.py
@@ -1,2 +1,0 @@
-from .a2c_acktr import A2C_ACKTR
-from .ppo import PPO
