+++ MarioWM/a2c_ppo_acktr/storage.py
@@ -1,26 +1,33 @@
 import torch
 from torch.utils.data.sampler import BatchSampler, SubsetRandomSampler
-from a2c_ppo_acktr.arguments import get_args
 
-args = get_args()
-trace_size = args.trace_size
-    
+'''
+Samples, transforms, and vectorizes data from the environment.
+'''
 
-    
+
 def _flatten_helper(T, N, _tensor):
     return _tensor.view(T * N, *_tensor.size()[2:])
 
 
 class RolloutStorage(object):
-    def __init__(self, num_steps, num_processes, obs_shape, action_space,
-                 recurrent_hidden_state_size):
+
+    def __init__(self, num_steps, num_processes, obs_shape, action_space, recurrent_hidden_state_size):
+
+        """Initialize Observations"""
+
         self.obs = torch.zeros(num_steps + 1, num_processes, *obs_shape)
         self.tobs = torch.zeros(num_steps + 1, num_processes, trace_size)
-        self.recurrent_hidden_states = torch.zeros(
-            num_steps + 1, num_processes, recurrent_hidden_state_size)
+        self.recurrent_hidden_states = torch.zeros(num_steps + 1, num_processes, recurrent_hidden_state_size)
+
+        '''Initialize Rewards'''
+
         self.rewards = torch.zeros(num_steps, num_processes, 1)
         self.value_preds = torch.zeros(num_steps + 1, num_processes, 1)
         self.returns = torch.zeros(num_steps + 1, num_processes, 1)
+
+        '''Initialize Actions'''
+
         self.action_log_probs = torch.zeros(num_steps, num_processes, 1)
         if action_space.__class__.__name__ == 'Discrete':
             action_shape = 1
@@ -29,12 +36,14 @@ class RolloutStorage(object):
         self.actions = torch.zeros(num_steps, num_processes, action_shape)
         if action_space.__class__.__name__ == 'Discrete':
             self.actions = self.actions.long()
-        self.masks = torch.ones(num_steps + 1, num_processes, 1)
 
+        '''Initialize Masks'''
+        self.masks = torch.ones(num_steps + 1, num_processes, 1)
         # Masks that indicate whether it's a true terminal state
         # or time limit end state
         self.bad_masks = torch.ones(num_steps + 1, num_processes, 1)
 
+        '''Initialize Steps'''
         self.num_steps = num_steps
         self.step = 0
 
@@ -43,7 +52,7 @@ class RolloutStorage(object):
         self.tobs = self.tobs.to(device)
 
         self.recurrent_hidden_states = self.recurrent_hidden_states.to(device)
-        self.rewards = self.rewards.to(device)
+        self.rewards = self.rewards.toe(device)
         self.value_preds = self.value_preds.to(device)
         self.returns = self.returns.to(device)
         self.action_log_probs = self.action_log_probs.to(device)
@@ -53,11 +62,17 @@ class RolloutStorage(object):
 
     def insert(self, obs, tobs, recurrent_hidden_states, actions, action_log_probs,
                value_preds, rewards, masks, bad_masks):
-        
+
+        '''
+        Insert prior observations at next step.
+        '''
         self.obs[self.step + 1].copy_(obs)
         self.tobs[self.step + 1].copy_(tobs)
-        self.recurrent_hidden_states[self.step +
-                                     1].copy_(recurrent_hidden_states)
+        self.recurrent_hidden_states[self.step + 1].copy_(recurrent_hidden_states)
+
+        '''
+        Insert prior actions at next step. 
+        '''
         self.actions[self.step].copy_(actions)
         self.action_log_probs[self.step].copy_(action_log_probs)
         self.value_preds[self.step].copy_(value_preds)
@@ -75,12 +90,7 @@ class RolloutStorage(object):
         self.masks[0].copy_(self.masks[-1])
         self.bad_masks[0].copy_(self.bad_masks[-1])
 
-    def compute_returns(self,
-                        next_value,
-                        use_gae,
-                        gamma,
-                        gae_lambda,
-                        use_proper_time_limits=True):
+    def compute_returns(self, next_value, use_gae, gamma, gae_lambda, use_proper_time_limits=True):
         if use_proper_time_limits:
             if use_gae:
                 self.value_preds[-1] = next_value
@@ -89,17 +99,16 @@ class RolloutStorage(object):
                     delta = self.rewards[step] + gamma * self.value_preds[
                         step + 1] * self.masks[step +
                                                1] - self.value_preds[step]
-                    #delta += 0.2 * abs(delta)
-                    gae = delta + gamma * gae_lambda * self.masks[step +
-                                                                  1] * gae
+                    # delta += 0.2 * abs(delta)
+                    gae = delta + gamma * gae_lambda * self.masks[step + 1] * gae
                     gae = gae * self.bad_masks[step + 1]
-                    self.returns[step] = gae + self.value_preds[step] 
+                    self.returns[step] = gae + self.value_preds[step]
             else:
                 self.returns[-1] = next_value
                 for step in reversed(range(self.rewards.size(0))):
                     self.returns[step] = (self.returns[step + 1] * \
-                        gamma * self.masks[step + 1] + self.rewards[step]) * self.bad_masks[step + 1] \
-                        + (1 - self.bad_masks[step + 1]) * self.value_preds[step]
+                                          gamma * self.masks[step + 1] + self.rewards[step]) * self.bad_masks[step + 1] \
+                                         + (1 - self.bad_masks[step + 1]) * self.value_preds[step]
         else:
             if use_gae:
                 self.value_preds[-1] = next_value
@@ -108,20 +117,19 @@ class RolloutStorage(object):
                     delta = self.rewards[step] + gamma * self.value_preds[
                         step + 1] * self.masks[step +
                                                1] - self.value_preds[step]
-                    #delta += 0.2 * abs(delta)
-                    gae = delta + gamma * gae_lambda * self.masks[step +
-                                                                  1] * gae
-                    self.returns[step] = gae + self.value_preds[step] 
+                    # delta += 0.2 * abs(delta)
+                    gae = delta + gamma * gae_lambda * self.masks[step + 1] * gae
+                    if use_proper_time_limits:
+                        gae = gae * self.bad_masks[step + 1]
+                    self.returns[step] = gae + self.value_preds[step]
             else:
                 self.returns[-1] = next_value
                 for step in reversed(range(self.rewards.size(0))):
+                    # this last calculation different with time limits
                     self.returns[step] = self.returns[step + 1] * \
-                        gamma * self.masks[step + 1] + self.rewards[step]
+                                         gamma * self.masks[step + 1] + self.rewards[step]
 
-    def feed_forward_generator(self,
-                               advantages,
-                               num_mini_batch=None,
-                               mini_batch_size=None):
+    def feed_forward_generator(self, advantages, num_mini_batch=None, mini_batch_size=None):
         num_steps, num_processes = self.rewards.size()[0:2]
         batch_size = num_processes * num_steps
 
@@ -133,10 +141,7 @@ class RolloutStorage(object):
                 "".format(num_processes, num_steps, num_processes * num_steps,
                           num_mini_batch))
             mini_batch_size = batch_size // num_mini_batch
-        sampler = BatchSampler(
-            SubsetRandomSampler(range(batch_size)),
-            mini_batch_size,
-            drop_last=True)
+        sampler = BatchSampler(SubsetRandomSampler(range(batch_size)), mini_batch_size, drop_last=True)
         for indices in sampler:
             obs_batch = self.obs[:-1].view(-1, *self.obs.size()[2:])[indices]
             tobs_batch = self.tobs[:-1].view(-1, *self.tobs.size()[2:])[indices]
@@ -156,7 +161,7 @@ class RolloutStorage(object):
                 adv_targ = advantages.view(-1, 1)[indices]
 
             yield obs_batch, tobs_batch, recurrent_hidden_states_batch, actions_batch, \
-                value_preds_batch, return_batch, masks_batch, old_action_log_probs_batch, adv_targ
+                  value_preds_batch, return_batch, masks_batch, old_action_log_probs_batch, adv_targ
 
     def recurrent_generator(self, advantages, num_mini_batch):
         num_processes = self.rewards.size(1)
@@ -168,7 +173,6 @@ class RolloutStorage(object):
         perm = torch.randperm(num_processes)
         for start_ind in range(0, num_processes, num_envs_per_batch):
             obs_batch = []
-            tobs_batch =[]
             recurrent_hidden_states_batch = []
             actions_batch = []
             value_preds_batch = []
@@ -180,8 +184,6 @@ class RolloutStorage(object):
             for offset in range(num_envs_per_batch):
                 ind = perm[start_ind + offset]
                 obs_batch.append(self.obs[:-1, ind])
-                #tobs_batch.append(self.obs[:-1, ind])
-                tobs_batch.append(self.tobs[:-1, ind])
                 recurrent_hidden_states_batch.append(
                     self.recurrent_hidden_states[0:1, ind])
                 actions_batch.append(self.actions[:, ind])
@@ -195,8 +197,6 @@ class RolloutStorage(object):
             T, N = self.num_steps, num_envs_per_batch
             # These are all tensors of size (T, N, -1)
             obs_batch = torch.stack(obs_batch, 1)
-            tobs_batch = torch.stack(tobs_batch, 1)
-
             actions_batch = torch.stack(actions_batch, 1)
             value_preds_batch = torch.stack(value_preds_batch, 1)
             return_batch = torch.stack(return_batch, 1)
@@ -211,15 +211,13 @@ class RolloutStorage(object):
 
             # Flatten the (T, N, ...) tensors to (T * N, ...)
             obs_batch = _flatten_helper(T, N, obs_batch)
-            tobs_batch = _flatten_helper(T, N, tobs_batch)
-
             actions_batch = _flatten_helper(T, N, actions_batch)
             value_preds_batch = _flatten_helper(T, N, value_preds_batch)
             return_batch = _flatten_helper(T, N, return_batch)
             masks_batch = _flatten_helper(T, N, masks_batch)
             old_action_log_probs_batch = _flatten_helper(T, N, \
-                    old_action_log_probs_batch)
+                                                         old_action_log_probs_batch)
             adv_targ = _flatten_helper(T, N, adv_targ)
 
-            yield obs_batch, tobs_batch, recurrent_hidden_states_batch, actions_batch, \
-                value_preds_batch, return_batch, masks_batch, old_action_log_probs_batch, adv_targ
+            yield obs_batch, recurrent_hidden_states_batch, actions_batch, \
+                  value_preds_batch, return_batch, masks_batch, old_action_log_probs_batch, adv_targ
