+++ MarioWM/a2c_ppo_acktr/utils/train_utils.py
@@ -1,0 +1,41 @@
+import torch.nn as nn
+
+
+def update_linear_schedule(optimizer, epoch, total_num_epochs, initial_lr):
+    """Linearly decrease learning rate"""
+    lr = initial_lr - (initial_lr * (epoch / float(total_num_epochs)))
+    for param_group in optimizer.param_groups:
+        param_group['lr'] = lr
+
+
+def init(module, weight_init, bias_init, gain=1):
+    weight_init(module.weight.data, gain=gain)
+    bias_init(module.bias.data)
+    return module
+
+
+class SplitBias(nn.Module):
+    def __init__(self, module):
+        super(SplitBias, self).__init__()
+        self.module = module
+        self.add_bias = AddBias(module.bias.data)
+        self.module.bias = None
+
+    def forward(self, input):
+        x = self.module(input)
+        x = self.add_bias(x)
+        return x
+
+
+class AddBias(nn.Module):
+    def __init__(self, bias):
+        super(AddBias, self).__init__()
+        self._bias = nn.Parameter(bias.unsqueeze(1))
+
+    def forward(self, x):
+        if x.dim() == 2:
+            bias = self._bias.t().view(1, -1)
+        else:
+            bias = self._bias.t().view(1, -1, 1, 1)
+
+        return x + bias
