+++ MarioWM/go.sh
