--- README.md
@@ -1,3 +1,0 @@
-# Framework for Weird Mario Machine Experimentation
-
-Explanation to follow.
