+++ RetroGym/retro/examples/trivial_random_agent.py
@@ -4,7 +4,7 @@ def main():
     env = retro.make(game='Airstriker-Genesis')
     obs = env.reset()
     while True:
-        obs, rew, done, info = env.step(env.action_space.sample())
+        obs, rew, done, info = env.step()
         env.render()
         if done:
             obs = env.reset()
