+++ RetroGym/retro/examples/trivial_random_agent_multiplayer.py
@@ -6,7 +6,7 @@ def main():
     while True:
         # action_space will by MultiBinary(16) now instead of MultiBinary(8)
         # the bottom half of the actions will be for player 1 and the top half for player 2
-        obs, rew, done, info = env.step(env.action_space.sample())
+        obs, rew, done, info = env.step()
         # rew will be a list of [player_1_rew, player_2_rew]
         # done and info will remain the same
         env.render()
